refactor(utils): move TextlintKeyPath from types to utils
