+++ packages/@textlint/kernel/package.json
@@ -35,6 +35,7 @@
     "@textlint/ast-node-types": "^4.2.2",
     "@textlint/ast-traverse": "^2.1.3",
     "@textlint/feature-flag": "^3.1.3",
+    "@textlint/utils": "^1.0.0",
     "@textlint/types": "^1.1.5",
     "@types/bluebird": "^3.5.18",
     "bluebird": "^3.5.1",
