+++ packages/textlint/package.json
@@ -48,6 +48,7 @@
     "watch": "tsc -p . --watch"
   },
   "dependencies": {
+    "@textlint/utils": "^1.0.0",
     "@textlint/ast-node-types": "^4.2.2",
     "@textlint/ast-traverse": "^2.1.3",
     "@textlint/feature-flag": "^3.1.3",
