+++ packages/utils/src/index.ts
@@ -1,0 +1,9 @@
+export {
+    normalizeTextlintPresetSubRuleKey,
+    normalizeTextlintRuleKey,
+    normalizeTextlintRulePresetKey,
+    normalizeTextlintPluginKey,
+    normalizeTextlintFilterRuleKey,
+    normalizeTextlintKeyPath
+} from "./KeyPath/TextlintKeyPath";
+export { TextlintPackageNamePrefix } from "./KeyPath/TextlintPackagePrefix";
